chat

+++ frontend/modules/chat/controllers/DefaultController.php
@@ -167,7 +167,7 @@ class DefaultController extends Controller
             {
 
                 
-                $messages = ChatModel::getRecentMessages($model->recipient_id, $this->module->numberLastMessages,$model->topic_id, $model->last_message_id);
+                $messages = ChatModel::getRefreshMessages($model->recipient_id, $this->module->numberLastMessages,$model->topic_id, $model->last_message_id);
                 $result = json_encode($messages);
 
                 if($messages){
@@ -205,6 +205,22 @@ class DefaultController extends Controller
         $chat_id = Yii::$app->request->post('cid');
 
         $model = ChatModel::findOne($chat_id);
+
+        $new = new ChatModel();
+        $new->is_deleted = $chat_id;
+        $new->topic_id = $model->topic_id;
+        $new->sender_id = $model->sender_id;
+        $new->recipient_id = $model->recipient_id;
+        $new->time = time();
+        $new->rfc822 = date(DATE_RFC822,$new->time);
+        $new->is_read = 1;
+
+        if(!$new->save()){
+            return json_encode($new->errors);
+        }
+
+        
+
         $model->delete();
         
         return $chat_id;
